+++ src/Collection/Factory/CollectionsManagerFactory.php
@@ -2,24 +2,21 @@
 
 namespace Svycka\Settings\Collection\Factory;
 
+use Interop\Container\ContainerInterface;
 use Svycka\Settings\Options\ModuleOptions;
 use Svycka\Settings\Collection\CollectionsManager;
-use Zend\ServiceManager\Config;
-use Zend\ServiceManager\FactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
 
 /**
  * @author Vytautas Stankus <svycka@gmail.com>
  * @license MIT
  */
-class CollectionsManagerFactory implements FactoryInterface
+final class CollectionsManagerFactory
 {
-    public function createService(ServiceLocatorInterface $serviceLocator)
+    public function __invoke(ContainerInterface $container)
     {
         /** @var ModuleOptions $options */
-        $options = $serviceLocator->get(ModuleOptions::class);
-        $serviceConfig = new Config($options->getSettingsManager());
+        $options = $container->get(ModuleOptions::class);
 
-        return new CollectionsManager($serviceConfig);
+        return new CollectionsManager($container, $options->getSettingsManager());
     }
 }