+++ src/Collection/Factory/SettingsCollectionAbstractFactory.php
@@ -2,6 +2,8 @@
 
 namespace Svycka\Settings\Collection\Factory;
 
+use Interop\Container\ContainerInterface;
+use Interop\Container\Exception\ContainerException;
 use Svycka\Settings\Collection\CollectionsManager;
 use Svycka\Settings\Collection\SettingsCollection;
 use Svycka\Settings\Options\CollectionOptions;
@@ -9,31 +11,27 @@ use Svycka\Settings\Options\ModuleOptions;
 use Svycka\Settings\Provider\OwnerProviderInterface;
 use Svycka\Settings\Storage\StorageAdapterInterface;
 use Svycka\Settings\Type\TypesManager;
-use Zend\ServiceManager\AbstractFactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use Zend\ServiceManager\Factory\AbstractFactoryInterface;
+use Zend\ServiceManager\Exception\ServiceNotCreatedException;
+use Zend\ServiceManager\Exception\ServiceNotFoundException;
 
 /**
  * @author Vytautas Stankus <svycka@gmail.com>
  * @license MIT
  */
-class SettingsCollectionAbstractFactory implements AbstractFactoryInterface
+final class SettingsCollectionAbstractFactory implements AbstractFactoryInterface
 {
     /**
-     * @var array
-     */
-    protected $config;
-
-    /**
-     * @param CollectionsManager|ServiceLocatorInterface $serviceLocator
-     * @param string                                     $name
-     * @param string                                     $requestedName
+     * Can the factory create an instance for the service?
      *
+     * @param  ContainerInterface $container
+     * @param  string $requestedName
      * @return bool
      */
-    public function canCreateServiceWithName(ServiceLocatorInterface $serviceLocator, $name, $requestedName)
+    public function canCreate(ContainerInterface $container, $requestedName)
     {
         /** @var CollectionsManager $config */
-        $config = $this->getConfig($serviceLocator);
+        $config = $container->get(ModuleOptions::class)->getCollections();
         if (empty($config)) {
             return false;
         }
@@ -42,48 +40,30 @@ class SettingsCollectionAbstractFactory implements AbstractFactoryInterface
     }
 
     /**
-     * @param CollectionsManager|ServiceLocatorInterface $serviceLocator
-     * @param string                                     $name
-     * @param string                                     $requestedName
+     * Create an object
      *
-     * @return SettingsCollection
+     * @param  ContainerInterface $container
+     * @param  string             $requestedName
+     * @param  null|array         $options
+     * @return object
+     * @throws ServiceNotFoundException if unable to resolve the service.
+     * @throws ServiceNotCreatedException if an exception is raised when
+     *     creating a service.
+     * @throws ContainerException if any other error occurs
      */
-    public function createServiceWithName(ServiceLocatorInterface $serviceLocator, $name, $requestedName)
+    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
     {
-        /** @var CollectionsManager $serviceLocator */
-        $serviceManager = $serviceLocator->getServiceLocator();
-
-        $config = new CollectionOptions($this->config[$requestedName]);
+        $collectionsConfig = $container->get(ModuleOptions::class)->getCollections();
+        $config = new CollectionOptions($collectionsConfig[$requestedName]);
         $config->setName($requestedName);
+
         /** @var StorageAdapterInterface $storage */
-        $storage = $serviceManager->get($config->getStorageAdapter());
+        $storage = $container->get($config->getStorageAdapter());
         /** @var OwnerProviderInterface $owner_provider */
-        $owner_provider = $serviceManager->get($config->getOwnerProvider());
+        $owner_provider = $container->get($config->getOwnerProvider());
         /** @var TypesManager $typesManager */
-        $typesManager = $serviceManager->get(TypesManager::class);
+        $typesManager = $container->get(TypesManager::class);
 
         return new SettingsCollection($config, $storage, $owner_provider, $typesManager);
-    }
-
-
-    /**
-     * Retrieve configuration for settings collection, if any
-     *
-     * @param  CollectionsManager $collectionsManager
-     *
-     * @return array
-     */
-    protected function getConfig(CollectionsManager $collectionsManager)
-    {
-        if ($this->config !== null) {
-            return $this->config;
-        }
-
-        /** @var ModuleOptions $options */
-        $options = $collectionsManager->getServiceLocator()->get(ModuleOptions::class);
-        $config  = $options->getCollections();
-
-        $this->config = $config;
-        return $this->config;
     }
 }