+++ src/Type/TypesManager.php
@@ -3,7 +3,8 @@
 namespace Svycka\Settings\Type;
 
 use Zend\ServiceManager\AbstractPluginManager;
-use Zend\ServiceManager\Exception\RuntimeException;
+use Zend\ServiceManager\Exception\InvalidServiceException;
+use Zend\ServiceManager\Factory\InvokableFactory;
 
 /**
  * Class TypesManager
@@ -15,29 +16,28 @@ use Zend\ServiceManager\Exception\RuntimeException;
  * @author Vytautas Stankus <svycka@gmail.com>
  * @license MIT
  */
-class TypesManager extends AbstractPluginManager
+final class TypesManager extends AbstractPluginManager
 {
     /**
      * @var array
      */
-    protected $invokableClasses = [
+    protected $factories = [
+        InArrayType::class => InvokableFactory::class,
+        RegexType::class => InvokableFactory::class,
+        IntegerType::class => InvokableFactory::class,
+        FloatType::class => InvokableFactory::class,
+        StringType::class => InvokableFactory::class,
+    ];
+
+    /**
+     * @var array
+     */
+    protected $aliases = [
         'inarray' => InArrayType::class,
         'regex'   => RegexType::class,
         'integer' => IntegerType::class,
         'float' => FloatType::class,
         'string' => StringType::class,
-    ];
-
-    /**
-     * @workaround: for https://github.com/zendframework/zend-servicemanager/issues/50
-     * @var array
-     */
-    protected $aliases = [
-        'svyckasettingstypeinarraytype' => 'inarray',
-        'svyckasettingstyperegextype'   => 'regex',
-        'svyckasettingstypeintegertype' => 'integer',
-        'svyckasettingstypefloattype' => 'float',
-        'svyckasettingstypestringtype' => 'string',
     ];
 
     /**
@@ -57,37 +57,15 @@ class TypesManager extends AbstractPluginManager
      * @return void
      * @throws \RuntimeException if invalid
      */
-    public function validatePlugin($type)
+    public function validate($type)
     {
         if ($type instanceof SettingTypeInterface) {
             return; // we're okay
         }
-        throw new RuntimeException(sprintf(
+        throw new InvalidServiceException(sprintf(
             'SettingType of type "%s" is invalid; must implement "%s"',
             (is_object($type) ? get_class($type) : gettype($type)),
             SettingTypeInterface::class
         ));
-    }
-
-    /**
-     * Override setInvokableClass().
-     *
-     * Performs normal operation, but also auto-aliases the class name to the
-     * service name. This ensures that providing the FQCN does not trigger an
-     * abstract factory later.
-     *
-     * @param  string       $name
-     * @param  string       $invokableClass
-     * @param  null|bool    $shared
-     * @return TypesManager
-     */
-    public function setInvokableClass($name, $invokableClass, $shared = null)
-    {
-        parent::setInvokableClass($name, $invokableClass, $shared);
-        if ($name != $invokableClass) {
-            $this->setAlias($invokableClass, $name);
-        }
-
-        return $this;
     }
 }